work in progress
Former-commit-id: b7122a0341bc0eef9947a8e2587be1eda73d0023
Former-commit-id: 9d3944cc5769bba8c4edb6c1e538732be4c6ba1f

+++ src/thing_private.h
@@ -79,7 +79,7 @@ typedef struct thing_ {
     /*
      * How many and of what we are carrying.
      */
-    uint16_t carrying[THING_MAX];
+    uint8_t carrying[THING_MAX];
 
     /*
      * Grid coordinates.
